LayoutTests/http/tests/loading/simple-subframe-expected.txt

main frame - didFinishDocumentLoadForFrame
frame "f1" - didStartProvisionalLoadForFrame
frame "f1" - didCommitLoadForFrame
frame "f1" - didFinishDocumentLoadForFrame
frame "f1" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "f1" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
This is a test of load callbacks. It is only useful inside the regression test tool.
